+++ app/src/main/java/com/elementary/tasks/module/KoinModule.kt
@@ -17,7 +17,6 @@ import com.elementary.tasks.module.featuresettings.TroubleshootingCacheUtilImpl
 import com.elementary.tasks.module.locationapi.LocationTrackingApiImpl
 import com.elementary.tasks.module.logicreminder.AppReminderNotifier
 import com.elementary.tasks.module.logicreminder.ReminderPreferencesImpl
-import com.elementary.tasks.reminder.scheduling.usecase.SnoozeReminderUseCaseImpl
 import com.elementary.tasks.module.logicschedule.SchedulePreferencesImpl
 import com.elementary.tasks.module.platform.BuildInfoImpl
 import com.elementary.tasks.module.platform.DateTimePreferencesImpl
@@ -49,7 +48,6 @@ import com.github.naz013.feature.settings.troubleshooting.TroubleshootingCacheUt
 import com.github.naz013.location.LocationTrackingApi
 import com.github.naz013.logic.reminder.ReminderNotifier
 import com.github.naz013.logic.reminder.ReminderPreferences
-import com.github.naz013.logic.reminder.usecase.SnoozeReminderUseCase
 import com.github.naz013.logic.schedule.SchedulePreferences
 import com.github.naz013.sync.SyncDataConverter
 import com.github.naz013.ui.common.font.FontApi
@@ -106,7 +104,6 @@ val libModule = module {
   // logic reminder
   factory { ReminderPreferencesImpl(get()) as ReminderPreferences }
   factory { AppReminderNotifier(get(), get(), get(), get()) as ReminderNotifier }
-  factory { SnoozeReminderUseCaseImpl(get(), get(), get(), get(), get(), get()) as SnoozeReminderUseCase }
 
   // location api
   factory { LocationTrackingApiImpl(get()) as LocationTrackingApi }
