+++ logic/logic-reminder/src/main/kotlin/com/github/naz013/logic/reminder/usecase/SnoozeReminderUseCaseImpl.kt
@@ -1,14 +1,11 @@
-package com.elementary.tasks.reminder.scheduling.usecase
+package com.github.naz013.logic.reminder.usecase
 
-import com.github.naz013.feature.workflow.WorkflowTriggerRunner
+import com.github.naz013.logic.reminder.ReminderWorkflowTrigger
 import com.github.naz013.domain.reminder.v2.ReminderV2
 import com.github.naz013.domain.sync.SyncState
 import com.github.naz013.logging.Logger
 import com.github.naz013.logic.reminder.behavior.BehaviorStrategyResolverV2
 import com.github.naz013.logic.reminder.behavior.LocationBasedStrategyV2
-import com.github.naz013.logic.reminder.usecase.CompleteReminderUseCase
-import com.github.naz013.logic.reminder.usecase.SaveReminderUseCase
-import com.github.naz013.logic.reminder.usecase.SnoozeReminderUseCase
 import com.github.naz013.notification.NotificationApi
 import com.github.naz013.scheduler.JobSchedulerApi
 
@@ -28,7 +25,7 @@ class SnoozeReminderUseCaseImpl(
   private val completeReminderUseCase: CompleteReminderUseCase,
   private val saveReminderUseCase: SaveReminderUseCase,
   private val notificationApi: NotificationApi,
-  private val workflowTriggerRunner: WorkflowTriggerRunner,
+  private val reminderWorkflowTrigger: ReminderWorkflowTrigger,
 ) : SnoozeReminderUseCase {
   override suspend operator fun invoke(
     reminder: ReminderV2,
@@ -53,7 +50,7 @@ class SnoozeReminderUseCaseImpl(
       )
     saveReminderUseCase(reminder)
     jobScheduler.scheduleReminderDelay(timeInMinutes, reminder.uuId, reminder.uniqueId)
-    workflowTriggerRunner.onReminderSnoozed(reminder.uuId)
+    reminderWorkflowTrigger.onReminderSnoozed(reminder.uuId)
     Logger.i(TAG, "Snoozed reminder id=${reminder.uuId} for $timeInMinutes minutes")
     return reminder
   }
