+++ logic/logic-workflow/src/main/kotlin/com/github/naz013/logic/workflow/WorkflowActionDispatcher.kt
@@ -1,15 +1,13 @@
-package com.github.naz013.feature.workflow
+package com.github.naz013.logic.workflow
 
 import com.github.naz013.domain.workflow.WorkflowAction
 import com.github.naz013.logic.reminder.usecase.ActivateReminderUseCase
 import com.github.naz013.logic.reminder.usecase.CompleteReminderUseCase
-import com.github.naz013.logic.workflow.PendingWorkflowAction
 import com.github.naz013.repository.ReminderV2Repository
 
 /**
- * Finishes the two [WorkflowAction] variants [WorkflowEngine][com.github.naz013.logic.workflow.WorkflowEngine]
- * (a pure-JVM module) can't apply itself, since both need `logic-reminder`'s
- * [ActivateReminderUseCase]/[CompleteReminderUseCase], which `usecase:reminders` doesn't depend on.
+ * Finishes the two [WorkflowAction] variants [WorkflowEngine] (a pure-JVM engine) can't apply
+ * itself, since both need `logic-reminder`'s [ActivateReminderUseCase]/[CompleteReminderUseCase].
  */
 class WorkflowActionDispatcher(
   private val reminderV2Repository: ReminderV2Repository,
