+++ logic/logic-workflow/src/main/kotlin/com/github/naz013/logic/workflow/WorkflowTriggerRunner.kt
@@ -1,16 +1,17 @@
-package com.github.naz013.feature.workflow
+package com.github.naz013.logic.workflow
 
-import com.github.naz013.logic.workflow.WorkflowEngine
+import com.github.naz013.logic.reminder.ReminderWorkflowTrigger
 
 /**
  * Single entry point shared by the periodic polling tasks and every reactive trigger call site:
- * runs the relevant [WorkflowEngine] method(s), then dispatches any [PendingWorkflowAction][com.github.naz013.logic.workflow.PendingWorkflowAction]
- * it returns via [WorkflowActionDispatcher].
+ * runs the relevant [WorkflowEngine] method(s), then dispatches any [PendingWorkflowAction]
+ * it returns via [WorkflowActionDispatcher]. Also implements [ReminderWorkflowTrigger], the seam
+ * `logic-reminder` uses to trigger workflow evaluation without depending on `logic-workflow`.
  */
 class WorkflowTriggerRunner(
   private val workflowEngine: WorkflowEngine,
   private val workflowActionDispatcher: WorkflowActionDispatcher
-) {
+) : ReminderWorkflowTrigger {
   suspend fun runDailyPolling() {
     (workflowEngine.runAgeBasedRules() + workflowEngine.runGroupCompletionRules())
       .forEach { workflowActionDispatcher.dispatch(it) }
@@ -24,7 +25,7 @@ class WorkflowTriggerRunner(
     workflowEngine.runReminderCompletedRules(reminderId).forEach { workflowActionDispatcher.dispatch(it) }
   }
 
-  suspend fun onReminderSnoozed(reminderId: String) {
+  override suspend fun onReminderSnoozed(reminderId: String) {
     workflowEngine.runSnoozeCountRules(reminderId).forEach { workflowActionDispatcher.dispatch(it) }
   }
